KAS-3499 move tests to seperate spec file
